DefaultGCActivityCallback::doWork() should reschedule if GC is deferred. <https://webkit.org/b/129131>
Reviewed by Mark Hahnenberg.

Currently, DefaultGCActivityCallback::doWork() does not check if the GC
needs to be deferred before commencing. As a result, the GC may crash
and/or corrupt data because the VM is not in the consistent state needed
for the GC to run. With this fix, doWork() now checks if the GC is
supposed to be deferred and re-schedules if needed. It only commences
with GC'ing when it's safe to do so.

* runtime/GCActivityCallback.cpp:
(JSC::DefaultGCActivityCallback::doWork):



Identifier: 147171@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@164463 268f45cc-cd09-0410-ab3c-d52691b4dbfc