[CSS Regions] Relative positioned elements overflowing the region do not get painted into the next tile https://bugs.webkit.org/show_bug.cgi?id=129254
Reviewed by Antti Koivisto.

Source/WebCore:

The painting of the region's layer should not be aborted so early if the region's fragment shouldn't be painted
because that would prevent all the layer's children from being painted.

Another problem this patch addresses is that clipping should also be performed when the clip rect is empty,
which is what happens when painting in a tile in which the flowed element would normally be painted
if it wasn't clipped by the region.

Test: fast/regions/content-relative-next-tile.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::paintFlowThreadIfRegionForFragments):

LayoutTests:

Added test for the painting across tiles of relative positioned elements flowed into regions.

* fast/regions/content-relative-next-tile-expected.html: Added.
* fast/regions/content-relative-next-tile.html: Added.


Identifier: 147308@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@164600 268f45cc-cd09-0410-ab3c-d52691b4dbfc