REGRESSION(r162947): Document::topDocument() returns an incorrect reference for cached Documents https://bugs.webkit.org/show_bug.cgi?id=128175
Reviewed by Antti Koivisto.

* dom/Document.cpp:
(WebCore::Document::topDocument): Fall back to pre-r162947 way of determining the top document
when the Document is in page cache or is in the middle of having its render tree destroyed.
In the first case, the determined top document is actually the document currently loaded in the
Frame to which the cached document is still connected, which is obviously not desired. In the
second case the top document is similarly incorrectly deduced, leading to non-deletion of the
proper top document's AXObjectCache. Because of this AccessibilityRenderObjects are not detached
which results in assertions in RenderObject destructor where the objects are found to be still
flagged as in use by the AX cache.


Identifier: 147412@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@164718 268f45cc-cd09-0410-ab3c-d52691b4dbfc