Ranges given to createMarkupInternal might not be canonicalized https://bugs.webkit.org/show_bug.cgi?id=129408
Reviewed by Ryosuke Niwa.

When WebKit clients (such as Mail) call WKWebArchiveCreateFromRange() with a range,
that range is not necessarily canonicalized. Therefore, comparing it for equality
with a canonicalized range might return a false negative. Instead, we should compare
canonicalized ranges.

No new tests because this codepath will only be reached via SPI.

* editing/markup.cpp:
(WebCore::createMarkupInternal):

Identifier: 147457@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@164763 268f45cc-cd09-0410-ab3c-d52691b4dbfc