Make Start Timeline Recording in the Develop menu show the Timeline view again.
This also fixes a noticeable delay between showing the sidebar and the default Timeline
view when initially opening the Web Inspector to the Timeline.

https://bugs.webkit.org/show_bug.cgi?id=129545

Reviewed by Joseph Pecoraro.

* UserInterface/Base/Main.js:
(WebInspector.contentLoaded): Call TimelineSidebarPanel.initialize to prime the content view.
This is needed so the view is ready in this run loop cycle, where it was delayed before.
* UserInterface/Protocol/InspectorFrontendAPI.js:
(InspectorFrontendAPI.setTimelineProfilingEnabled): Call TimelineSidebarPanel.showTimelineOverview.
* UserInterface/Views/TimelineSidebarPanel.js:
(WebInspector.TimelineSidebarPanel.prototype.initialize): Added.


Identifier: 147601@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@164921 268f45cc-cd09-0410-ab3c-d52691b4dbfc