Split TextIteratorBehavior into a separate header https://bugs.webkit.org/show_bug.cgi?id=129578
Reviewed by Sam Weinig.

Source/WebCore:

This is in preparation to greatly cut down on includes of TextIterator.h.

* GNUmakefile.list.am: Added new header.
* WebCore.vcxproj/WebCore.vcxproj: Ditto.
* WebCore.vcxproj/WebCore.vcxproj.filters: Ditto.
* WebCore.xcodeproj/project.pbxproj: Ditto.

 * editing/FindOptions.h: Changed typedef to unsigned char;
there is no reason to store these in a 32-bit integer.

 * editing/TextIterator.h: Added include of TextIteratorBehavior.h
and moved comments about the meanings of flag bits to their definition.

* editing/TextIteratorBehavior.h: Added.

* WebCore.exp.in: Updated for type changes.

Tools:

* Scripts/sort-export-file: Added special cases to the script so you can just paste link errors
from WebKit in and the script will add them in as unconditional exports. This makes the "add"
task easy, but the remove is still a bit of a pain.


Identifier: 147637@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@164957 268f45cc-cd09-0410-ab3c-d52691b4dbfc